+++ bank-feeds/src/sdk/sdk.ts
@@ -57,8 +57,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "2.2.2";
-    genVersion = "2.131.1";
+    sdkVersion = "2.3.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 2.3.0 2.139.1 3.0.0 @codat/bank-feeds";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);